rebase to 3.2 and opt code
