+++ dubbo-rpc/dubbo-rpc-api/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.rpc.Filter
@@ -1,5 +1,6 @@
 echo=org.apache.dubbo.rpc.filter.EchoFilter
 generic=org.apache.dubbo.rpc.filter.GenericFilter
+omnipotent=org.apache.dubbo.rpc.filter.OmnipotentFilter
 genericimpl=org.apache.dubbo.rpc.filter.GenericImplFilter
 token=org.apache.dubbo.rpc.filter.TokenFilter
 accesslog=org.apache.dubbo.rpc.filter.AccessLogFilter
