+++ dubbo-rpc/dubbo-rpc-dubbo/src/main/java/org/apache/dubbo/rpc/protocol/dubbo/ServiceNotFoundExceptionProcessor.java
@@ -23,17 +23,17 @@ import org.apache.dubbo.remoting.ServiceNotFoundException;
 import org.apache.dubbo.remoting.exchange.ExchangeChannel;
 import org.apache.dubbo.remoting.exchange.Request;
 import org.apache.dubbo.rpc.service.OmnipotentService;
-import org.apache.dubbo.rpc.support.RpcUtils;
 
-import static org.apache.dubbo.common.constants.CommonConstants.PATH_KEY;
-import static org.apache.dubbo.common.constants.CommonConstants.INTERFACE_KEY;
-import static org.apache.dubbo.common.constants.CommonConstants.VERSION_KEY;
 import static org.apache.dubbo.common.constants.CommonConstants.DEFAULT_VERSION;
+import static org.apache.dubbo.common.constants.CommonConstants.INTERFACE_KEY;
 import static org.apache.dubbo.common.constants.CommonConstants.METHOD_KEY;
-import static org.apache.dubbo.common.constants.CommonConstants.$INVOKE_ASYNC;
-import static org.apache.dubbo.common.constants.CommonConstants.$INVOKE;
+import static org.apache.dubbo.common.constants.CommonConstants.PATH_KEY;
+import static org.apache.dubbo.common.constants.CommonConstants.VERSION_KEY;
+import static org.apache.dubbo.common.constants.OmnipotentCommonConstants.$INVOKE_OMN;
 import static org.apache.dubbo.common.constants.OmnipotentCommonConstants.GENERIC_PARAMETER_TYPES;
-import static org.apache.dubbo.rpc.Constants.GENERIC_KEY;
+import static org.apache.dubbo.common.constants.OmnipotentCommonConstants.ORIGIN_METHOD_KEY;
+import static org.apache.dubbo.common.constants.OmnipotentCommonConstants.ORIGIN_PATH_KEY;
+import static org.apache.dubbo.common.constants.OmnipotentCommonConstants.ORIGIN_VERSION_KEY;
 
 /**
  * <p>The default exception handler interrupts the return of the client error message by throwing a {@link RetryHandleException},
@@ -41,15 +41,11 @@ import static org.apache.dubbo.rpc.Constants.GENERIC_KEY;
  * <p>Then handle the {@link ServiceNotFoundException} through server-side generalization {@link OmnipotentService}
  * (different from the original generalization, which can accept any interface)
  *
- * @since 3.1.4
+ * @since 3.2.0
  */
 public class ServiceNotFoundExceptionProcessor implements ExceptionProcessor {
 
     private static final String DEFAULT_OMNIPOTENT_SERVICE = OmnipotentService.class.getName();
-    private static final String ORIGIN_PATH_KEY = "originPath";
-    private static final String ORIGIN_VERSION_KEY = "originVersion";
-    private static final String ORIGIN_METHOD_KEY = "originMethod";
-
 
     @Override
     public boolean shouldReturnError(Throwable data) {
@@ -69,22 +65,16 @@ public class ServiceNotFoundExceptionProcessor implements ExceptionProcessor {
         invocation.setAttachment(PATH_KEY, DEFAULT_OMNIPOTENT_SERVICE);
         invocation.setAttachment(INTERFACE_KEY, DEFAULT_OMNIPOTENT_SERVICE);
 
+        // version
         invocation.setAttachment(ORIGIN_VERSION_KEY, invocation.getAttachment(VERSION_KEY));
         invocation.setAttachment(VERSION_KEY, DEFAULT_VERSION);
 
+        // method
         invocation.setAttachment(ORIGIN_METHOD_KEY, invocation.getMethodName());
-        if (RpcUtils.isReturnTypeFuture(invocation)) {
-            invocation.setAttachment(METHOD_KEY, $INVOKE_ASYNC);
-            invocation.setMethodName($INVOKE_ASYNC);
-        } else {
-            invocation.setAttachment(METHOD_KEY, $INVOKE);
-            invocation.setMethodName($INVOKE);
-        }
-
+        invocation.setAttachment(METHOD_KEY, $INVOKE_OMN);
+        invocation.setMethodName($INVOKE_OMN);
         invocation.setParameterTypes(GENERIC_PARAMETER_TYPES);
 
-        invocation.setAttachment(
-                GENERIC_KEY, Boolean.TRUE.toString());
         invocation.resetHasDecoded();
 
         throw new RetryHandleException(channel, "Service not found:" + invocation.getAttachment(ORIGIN_PATH_KEY) + ", " + invocation.getAttachment(ORIGIN_METHOD_KEY));
